backend/requirements.txt
google-adk>=0.5.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
twilio>=9.0.0
google-cloud-firestore>=2.16.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
